requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
cfgv==3.4.0
click==8.1.7
cloudinary==1.37.0
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.103.2
filelock==3.13.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
identify==2.5.32
idna==3.6
iniconfig==2.0.0
mongoengine==0.27.0
mongomock==4.1.2
nodeenv==1.8.0
packaging==23.2
passlib==1.7.4
platformdirs==4.0.0
pluggy==1.3.0
pre-commit==3.5.0
pyasn1==0.5.1
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.10.1
pymongo==4.6.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
PyYAML==6.0.1
rsa==4.9
ruff==0.1.7
sentinels==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.23.2
virtualenv==20.25.0
